Move Event serialization logic to EventSerializer.
